backend/requirements.txt
# Core Framework
Flask==3.0.0
flask-cors==4.0.0

# Security
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# Input Validation
validators==0.22.0
bleach==6.1.0
python-dotenv==1.0.0

# ML & Data Processing
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Network & URL Processing
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0
urllib3==2.1.0

# SSL Certificate Handling
pyOpenSSL==23.3.0
certifi==2023.11.17

# Utilities
python-dateutil==2.8.2
